Correct failed test cases

+++ keras/src/layers/preprocessing/image_preprocessing/mix_up.py
@@ -58,7 +58,7 @@ class MixUp(BaseImagePreprocessingLayer):
             seed = self._get_seed_generator(self.backend._backend)
 
         permutation_order = self.backend.random.shuffle(
-            self.backend.numpy.arange(0, batch_size), seed=seed
+            self.backend.numpy.arange(0, batch_size, dtype="int64"), seed=seed
         )
 
         mix_weight = keras.src.random.random.beta(
